requirements.txt
django==3.2.6
django-bootstrap-v5==1.0.5
django-filter==2.4.0